Extracted methods into own API file.  Updated JsDoc
